requirements.txt
akshare==1.15.34
backports-datetime-fromisoformat==2.0.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.0
crypto==1.4.1
decorator==5.1.1
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
future==1.0.0
html5lib==1.1
idna==3.10
jsonpath==0.82.2
lxml==5.3.0
m3u8==6.0.0
mini-racer==0.12.4
Naked==0.1.32
numpy==2.2.0
openpyxl==3.1.5
pandas==2.2.3
pycryptodome==3.21.0
pycryptodomex==3.21.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
six==1.17.0
soupsieve==2.6
tabulate==0.9.0
tenacity==9.0.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
xlrd==2.0.1
pillow==11.0.0
